chore: split the 'DeeplinkManager' class into multiple chunks
