+++ app/core/DeeplinkManager/ParseManager/handleDappProtocol.ts
@@ -1,0 +1,16 @@
+import { DeeplinkManager } from '../DeeplinkManager';
+import extractURLParams from './extractURLParams';
+
+export function handleDappProtocol(
+  instance: DeeplinkManager,
+  handled: () => void,
+  urlObj: ReturnType<typeof extractURLParams>['urlObj'],
+  browserCallBack: (url: string) => void,
+) {
+  // Enforce https
+  handled();
+  urlObj.set('protocol', 'https:');
+  instance._handleBrowserUrl(urlObj.href, browserCallBack);
+}
+
+export default handleDappProtocol;
