+++ app/core/DeeplinkManager/handlers/handleBrowserUrl.ts
@@ -1,0 +1,25 @@
+import Routes from '../../../constants/navigation/Routes';
+import { InteractionManager } from 'react-native';
+import { DeeplinkManager } from '../DeeplinkManager';
+
+function handleBrowserUrl(
+  instance: DeeplinkManager,
+  url: string,
+  callback: (url: string) => void,
+) {
+  InteractionManager.runAfterInteractions(() => {
+    if (callback) {
+      callback(url);
+    } else {
+      instance.navigation.navigate(Routes.BROWSER.HOME, {
+        screen: Routes.BROWSER.VIEW,
+        params: {
+          newTabUrl: url,
+          timestamp: Date.now(),
+        },
+      });
+    }
+  });
+}
+
+export default handleBrowserUrl;
